requirements.txt
blessings==1.7
bpython==0.18
certifi==2020.6.20
chardet==3.0.4
curtsies==0.3.4
DateTime==4.3
Django==2.2.3
greenlet==0.4.17
idna==2.10
mysql-connector-python==8.0.21
pkg-resources==0.0.0
protobuf==3.13.0
Pygments==2.7.2
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
wcwidth==0.2.5
zope.interface==5.1.2
django-tinymce==3.0.2
Pillow==7.1.2